+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Head
+# 7.0.0 - 2016-11-09
 
 - Added: `template` option instead `css`, `cssFormat`, `srcCssTemplate`.
 - Added: `destStyles` options instead `destCssTemplate`.